requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
phonenumbers==9.0.10
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
smsaero_api==3.0.0
smsmobileapi==2.0.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
